Moving explode out of reshape and some test cleanup
